requirements.txt
asgiref==3.5.0
Django==4.0.2
django-ckeditor==6.5.1
django-js-asset==2.0.0
Pillow==9.5.0
psycopg2-binary==2.9.3
sqlparse==0.4.2
tzdata==2021.5
